Move BRS_X05 to BRS_X03.FailingOrchestrationInstanceExample
