+++ source/ProcessManager.Core.Tests/Unit/Infrastructure/Extensions/DependencyInjection/ProcessManagerExtensionsTests.cs
@@ -17,6 +17,7 @@ using Energinet.DataHub.ProcessManager.Core.Application.Orchestration;
 using Energinet.DataHub.ProcessManager.Core.Application.Registration;
 using Energinet.DataHub.ProcessManager.Core.Infrastructure.Extensions.DependencyInjection;
 using Energinet.DataHub.ProcessManager.Example.Orchestrations.Processes.BRS_X01.InputExample.V1.Options;
+using Energinet.DataHub.ProcessManager.Example.Orchestrations.Processes.BRS_X03.FailingOrchestrationInstanceExample;
 using Energinet.DataHub.ProcessManager.Shared.Tests.Fixtures.Extensions;
 using FluentAssertions;
 using FluentAssertions.Execution;
@@ -50,7 +51,7 @@ public class ProcessManagerExtensionsTests
             typeof(Example.Orchestrations.Processes.BRS_X02.ActorRequestProcessExample.OrchestrationDescriptionBuilder),
             typeof(Example.Orchestrations.Processes.BRS_X03.OrchestrationDescriptionBreakingChanges.BreakingChangesOrchestrationDescriptionBuilder),
             typeof(Example.Orchestrations.Processes.BRS_X03.OrchestrationDescriptionBreakingChanges.UnderDevelopmentOrchestrationDescriptionBuilder),
-            typeof(Example.Orchestrations.Processes.BRS_X05_FailingOrchestrationInstanceExample.OrchestrationDescriptionBuilder),
+            typeof(OrchestrationDescriptionBuilder),
         };
 
         // Act
